refactor: split precedence.rs into a folder module
1015-line precedence.rs had three cohesive groups: env-var wire format,
diff output types with rendering, and the Precedence builder engine.
Split into separate files under crates/am/src/precedence/ so each file
holds one responsibility and stays a comfortable read.

- precedence/mod.rs       — module declarations + re-exports only
- precedence/env_state.rs — AliasWithHash + AliasWithHashList (~193 lines)
- precedence/diff.rs      — EntryKind, EffectiveEntry, PrecedenceDiff
                            with change_summary + render (~194 lines)
- precedence/engine.rs    — Precedence builder + resolve + hashing helpers
                            (~653 lines)

Public API is unchanged. All external uses of
amoxide::precedence::{Precedence, PrecedenceDiff, EntryKind,
EffectiveEntry, AliasWithHash, AliasWithHashList} still resolve. No
signatures, behavior, or test count changed.
