+++ crates/am/src/precedence/engine.rs
@@ -1,165 +1,10 @@
 use std::collections::{BTreeMap, BTreeSet, HashSet};
-use std::fmt;
 
 use crate::alias::{AliasSet, TomlAlias};
-use crate::env_vars;
-use crate::shell::ShellAdapter;
-use crate::subcommand::{SubcommandEntry, SubcommandSet};
+use crate::subcommand::SubcommandSet;
 
-/// One entry in the `_AM_ALIASES` / `_AM_SUBCOMMANDS` env var, in the
-/// `"name|hash"` format (or a legacy bare `"name"` with no hash).
-///
-/// `hash = None` means the shell reloaded from an older amoxide that only
-/// tracked names; the diff treats such entries as "always differs" so they
-/// get reloaded on the next sync.
-#[derive(Debug, Clone, PartialEq)]
-pub struct AliasWithHash {
-    name: String,
-    hash: Option<String>,
-}
-
-impl AliasWithHash {
-    pub fn new(name: impl Into<String>, hash: Option<String>) -> Self {
-        Self {
-            name: name.into(),
-            hash,
-        }
-    }
-
-    pub fn name(&self) -> &str {
-        &self.name
-    }
-
-    pub fn hash(&self) -> Option<&str> {
-        self.hash.as_deref()
-    }
-
-    /// Parse one `"name|hash"` (or bare `"name"`) token. Returns `None` when
-    /// the name segment is empty — callers skip such entries silently.
-    pub fn parse(token: &str) -> Option<Self> {
-        match token.split_once('|') {
-            Some((name, hash)) if !name.is_empty() => Some(Self {
-                name: name.to_string(),
-                hash: Some(hash.to_string()),
-            }),
-            Some(_) => None, // empty name before '|'
-            None if token.is_empty() => None,
-            None => Some(Self {
-                name: token.to_string(),
-                hash: None,
-            }),
-        }
-    }
-}
-
-impl fmt::Display for AliasWithHash {
-    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        match &self.hash {
-            Some(h) => write!(f, "{}|{}", self.name, h),
-            None => write!(f, "{}", self.name),
-        }
-    }
-}
-
-/// A comma-separated list of [`AliasWithHash`] entries — the on-the-wire
-/// format of `_AM_ALIASES` and `_AM_SUBCOMMANDS`.
-///
-/// Owns round-trip parsing and rendering so no other module has to know
-/// about the `"name|hash,name|hash,..."` layout.
-#[derive(Debug, Default, Clone, PartialEq)]
-pub struct AliasWithHashList(Vec<AliasWithHash>);
-
-impl AliasWithHashList {
-    pub fn new() -> Self {
-        Self::default()
-    }
-
-    pub fn push(&mut self, entry: AliasWithHash) {
-        self.0.push(entry);
-    }
-
-    pub fn is_empty(&self) -> bool {
-        self.0.is_empty()
-    }
-
-    pub fn iter(&self) -> std::slice::Iter<'_, AliasWithHash> {
-        self.0.iter()
-    }
-
-    /// Parse an `_AM_ALIASES` / `_AM_SUBCOMMANDS` value. `None` or empty
-    /// string yields an empty list; malformed tokens are skipped.
-    pub fn parse(raw: Option<&str>) -> Self {
-        let Some(s) = raw.filter(|s| !s.is_empty()) else {
-            return Self::new();
-        };
-        Self(s.split(',').filter_map(AliasWithHash::parse).collect())
-    }
-}
-
-impl fmt::Display for AliasWithHashList {
-    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        for (i, entry) in self.0.iter().enumerate() {
-            if i > 0 {
-                f.write_str(",")?;
-            }
-            write!(f, "{entry}")?;
-        }
-        Ok(())
-    }
-}
-
-impl FromIterator<AliasWithHash> for AliasWithHashList {
-    fn from_iter<I: IntoIterator<Item = AliasWithHash>>(iter: I) -> Self {
-        Self(iter.into_iter().collect())
-    }
-}
-
-impl<'a> IntoIterator for &'a AliasWithHashList {
-    type Item = &'a AliasWithHash;
-    type IntoIter = std::slice::Iter<'a, AliasWithHash>;
-    fn into_iter(self) -> Self::IntoIter {
-        self.0.iter()
-    }
-}
-
-impl IntoIterator for AliasWithHashList {
-    type Item = AliasWithHash;
-    type IntoIter = std::vec::IntoIter<AliasWithHash>;
-    fn into_iter(self) -> Self::IntoIter {
-        self.0.into_iter()
-    }
-}
-
-#[derive(Debug, Clone, PartialEq)]
-pub enum EntryKind {
-    Alias(TomlAlias),
-    SubcommandWrapper {
-        program: String,
-        entries: Vec<SubcommandEntry>,
-        base_cmd: Option<String>,
-    },
-    /// Per-key subcommand entry tracked in `_AM_SUBCOMMANDS` for fine-grained
-    /// change detection. Never emitted as shell code — the program-level
-    /// `SubcommandWrapper` is the shell-visible unit.
-    SubcommandKey {
-        longs: Vec<String>,
-    },
-}
-
-#[derive(Debug, Clone, PartialEq)]
-pub struct EffectiveEntry {
-    pub name: String,
-    pub kind: EntryKind,
-    pub hash: String,
-}
-
-#[derive(Debug, Default, Clone, PartialEq)]
-pub struct PrecedenceDiff {
-    pub added: Vec<EffectiveEntry>,
-    pub changed: Vec<EffectiveEntry>,
-    pub removed: Vec<String>,
-    pub unchanged: Vec<EffectiveEntry>,
-}
+use super::diff::{EffectiveEntry, EntryKind, PrecedenceDiff};
+use super::env_state::AliasWithHashList;
 
 #[derive(Debug, Default)]
 pub struct Precedence {
@@ -413,175 +258,10 @@ impl Precedence {
     }
 }
 
-impl PrecedenceDiff {
-    /// Human-readable summary of what changed, suitable for echoing to the
-    /// user (e.g. `"am: aliases changed — 1 loaded: b | 1 unloaded: t"`).
-    ///
-    /// Returns `None` when nothing changed so callers can stay silent.
-    pub fn change_summary(&self) -> Option<String> {
-        let added: Vec<&str> = self.added.iter().map(|e| e.name.as_str()).collect();
-        let changed: Vec<&str> = self.changed.iter().map(|e| e.name.as_str()).collect();
-        let removed: Vec<&str> = self.removed.iter().map(|s| s.as_str()).collect();
-        let parts: Vec<String> = [
-            ("loaded", &added[..]),
-            ("updated", &changed[..]),
-            ("unloaded", &removed[..]),
-        ]
-        .iter()
-        .filter(|(_, names)| !names.is_empty())
-        .map(|(verb, names)| format!("{} {verb}: {}", names.len(), names.join(", ")))
-        .collect();
-        if parts.is_empty() {
-            None
-        } else {
-            Some(format!("am: aliases changed — {}", parts.join(" | ")))
-        }
-    }
-
-    /// Render this diff into shell code using the given adapter.
-    ///
-    /// Emission order:
-    ///   1. unload (removed + changed) — skipping subcommand-key names
-    ///      (they're tracking-only, not shell functions)
-    ///   2. load (added + changed)
-    ///   3. set `_AM_ALIASES` / `_AM_SUBCOMMANDS` to the union of added +
-    ///      changed + unchanged
-    pub fn render(&self, shell: &dyn ShellAdapter) -> String {
-        let mut lines: Vec<String> = Vec::new();
-
-        // 1. Unload
-        for name in &self.removed {
-            if name.contains(':') {
-                continue;
-            }
-            lines.push(shell.unalias(name));
-        }
-        for entry in &self.changed {
-            if matches!(entry.kind, EntryKind::SubcommandKey { .. }) {
-                continue;
-            }
-            if entry.name.contains(':') {
-                continue;
-            }
-            lines.push(shell.unalias(&entry.name));
-        }
-
-        // 2. Load (added + changed)
-        for entry in self.added.iter().chain(self.changed.iter()) {
-            match &entry.kind {
-                EntryKind::Alias(alias) => {
-                    lines.push(shell.alias(&alias.as_entry(&entry.name)));
-                }
-                EntryKind::SubcommandWrapper {
-                    program,
-                    entries,
-                    base_cmd,
-                } => {
-                    let cmd = base_cmd
-                        .clone()
-                        .unwrap_or_else(|| format!("command {program}"));
-                    lines.push(shell.subcommand_wrapper(program, &cmd, entries));
-                }
-                EntryKind::SubcommandKey { .. } => {}
-            }
-        }
-
-        // 3. Update tracking env vars
-        let mut alias_list = AliasWithHashList::new();
-        let mut sub_list = AliasWithHashList::new();
-        for e in self
-            .added
-            .iter()
-            .chain(self.changed.iter())
-            .chain(self.unchanged.iter())
-        {
-            let entry = AliasWithHash::new(&e.name, Some(e.hash.clone()));
-            match &e.kind {
-                EntryKind::SubcommandKey { .. } => sub_list.push(entry),
-                _ => alias_list.push(entry),
-            }
-        }
-
-        if !alias_list.is_empty() {
-            lines.push(shell.set_env(env_vars::AM_ALIASES, &alias_list.to_string()));
-        }
-        if !sub_list.is_empty() {
-            lines.push(shell.set_env(env_vars::AM_SUBCOMMANDS, &sub_list.to_string()));
-        }
-
-        lines.join("\n")
-    }
-}
-
 #[cfg(test)]
 mod tests {
     use super::*;
     use crate::alias::AliasName;
-
-    // ─── AliasWithHash / AliasWithHashList ──────────────────────────────
-
-    #[test]
-    fn alias_with_hash_parse_new_format() {
-        let e = AliasWithHash::parse("b|abc1234").unwrap();
-        assert_eq!(e.name(), "b");
-        assert_eq!(e.hash(), Some("abc1234"));
-    }
-
-    #[test]
-    fn alias_with_hash_parse_bare_name() {
-        let e = AliasWithHash::parse("t").unwrap();
-        assert_eq!(e.name(), "t");
-        assert_eq!(e.hash(), None);
-    }
-
-    #[test]
-    fn alias_with_hash_parse_empty_returns_none() {
-        assert!(AliasWithHash::parse("").is_none());
-        assert!(AliasWithHash::parse("|abc").is_none());
-    }
-
-    #[test]
-    fn alias_with_hash_display_roundtrip() {
-        assert_eq!(
-            AliasWithHash::new("b", Some("abc1234".into())).to_string(),
-            "b|abc1234"
-        );
-        assert_eq!(AliasWithHash::new("t", None).to_string(), "t");
-    }
-
-    #[test]
-    fn alias_with_hash_list_parse_and_render() {
-        let list = AliasWithHashList::parse(Some("b|abc1234,t|def5678"));
-        assert_eq!(list.iter().count(), 2);
-        assert_eq!(list.to_string(), "b|abc1234,t|def5678");
-    }
-
-    #[test]
-    fn alias_with_hash_list_parse_mixed_format() {
-        let list = AliasWithHashList::parse(Some("b|abc1234,t,gs|fed9876"));
-        let names: Vec<&str> = list.iter().map(|e| e.name()).collect();
-        assert_eq!(names, vec!["b", "t", "gs"]);
-        assert_eq!(list.iter().nth(1).unwrap().hash(), None);
-    }
-
-    #[test]
-    fn alias_with_hash_list_parse_empty_and_none() {
-        assert!(AliasWithHashList::parse(None).is_empty());
-        assert!(AliasWithHashList::parse(Some("")).is_empty());
-    }
-
-    #[test]
-    fn alias_with_hash_list_parse_skips_malformed_tokens() {
-        // Leading empty token and "|xxx" token get dropped silently.
-        let list = AliasWithHashList::parse(Some(",|xxx,b|abc1234"));
-        let names: Vec<&str> = list.iter().map(|e| e.name()).collect();
-        assert_eq!(names, vec!["b"]);
-    }
-
-    #[test]
-    fn alias_with_hash_list_display_empty() {
-        assert_eq!(AliasWithHashList::new().to_string(), "");
-    }
 
     #[test]
     fn empty_inputs_produce_empty_diff() {
@@ -969,47 +649,5 @@ mod tests {
             find(&diff.unchanged, "jj:ab").is_some(),
             "jj:ab entry itself is unchanged"
         );
-    }
-
-    use crate::config::ShellsTomlConfig;
-    use crate::shell::Shell;
-
-    #[test]
-    fn render_emits_unloads_then_loads_then_env() {
-        let cfg = ShellsTomlConfig::default();
-        let shell = Shell::Fish.as_shell(&cfg, Default::default(), Default::default());
-
-        // Previous shell state: `b|0000000,gone|aaa` ; new effective: `b|make build`.
-        let project = aset(&[("b", "make build")]);
-        let diff = Precedence::new()
-            .with_project(&project, &SubcommandSet::new())
-            .with_shell_state_from_env(Some("b|0000000,gone|aaa"), None)
-            .resolve();
-
-        let out = diff.render(shell.as_ref());
-        assert!(
-            out.contains("functions -e gone"),
-            "gone must be unloaded: {out}"
-        );
-        assert!(
-            out.contains("functions -e b"),
-            "changed b must be unloaded: {out}"
-        );
-        assert!(
-            out.contains("function b\n    make build $argv\nend"),
-            "b must be reloaded: {out}"
-        );
-        // env-var update must be the last section
-        let env_pos = out.find("_AM_ALIASES").expect("env update missing");
-        let fn_pos = out.find("function b").unwrap();
-        assert!(env_pos > fn_pos, "env update must come after loads");
-    }
-
-    #[test]
-    fn render_empty_diff_produces_empty_string() {
-        let cfg = ShellsTomlConfig::default();
-        let shell = Shell::Fish.as_shell(&cfg, Default::default(), Default::default());
-        let out = PrecedenceDiff::default().render(shell.as_ref());
-        assert!(out.is_empty());
     }
 }
